address review: move bagel SDE scheduler out of _shared into bagel dir; rename backend wire_dtype to weight_sync_dtype
