+++ unirl/distributed/weight_sync/lora/base.py
@@ -33,19 +33,22 @@ def _extract_canonical_lora(backend: Any, *, param_prefix: str, adapter_name: st
     tensor — a collective across the train process group — so the caller MUST run
     this on every train rank in lockstep (``BROADCAST``).
 
-    The wire dtype is the backend's FSDP compute dtype (``backend.wire_dtype``,
-    i.e. ``param_dtype``), NOT the trainable params' own dtype: under
-    ``master_dtype=fp32`` (the reward-collapse fix) the LoRA params are fp32, but
-    the rollout engine's vLLM punica kernel requires bf16/fp16 — so cast at the
-    all-gather. Falls back to ``None`` (keep dtype) for backends predating
-    ``wire_dtype`` (e.g. an all-bf16-master setup where no cast is needed).
+    The weight-sync dtype is the backend's FSDP compute dtype
+    (``backend.weight_sync_dtype``, i.e. ``param_dtype``), NOT the trainable
+    params' own dtype: under ``master_dtype=fp32`` (the reward-collapse fix) the
+    LoRA params are fp32, but the rollout engine's vLLM punica kernel requires
+    bf16/fp16 — so cast at the all-gather. Falls back to ``None`` (keep dtype) for
+    backends predating ``weight_sync_dtype`` (e.g. an all-bf16-master setup where
+    no cast is needed).
     """
     from unirl.distributed.weight_sync.payload import _peft_config_dict
     from unirl.utils.peft_merge import extract_lora_tensors
 
     model = backend.model
-    wire_dtype = getattr(backend, "wire_dtype", None)
-    lora_tensors = extract_lora_tensors(model, param_prefix=param_prefix, adapter_name=adapter_name, dtype=wire_dtype)
+    weight_sync_dtype = getattr(backend, "weight_sync_dtype", None)
+    lora_tensors = extract_lora_tensors(
+        model, param_prefix=param_prefix, adapter_name=adapter_name, dtype=weight_sync_dtype
+    )
     peft_config = _peft_config_dict(model, adapter_name)
     return lora_tensors, peft_config
